backend/requirements/base.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
openai==1.12.0
python-dotenv==1.0.0
alembic==1.13.1
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# RAG & Vector Database
qdrant-client==1.7.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# Text Processing
nltk==3.8.1
tiktoken==0.5.2
